Add MaxMind GeoIP test DB to use in the tests
Obtained from: https://github.com/maxmind/MaxMind-DB/blob/689bbf927360cee2285fd472fc1eb6671f335979/test-data/MaxMind-DB-test-ipv4-24.mmdb

The database is tiny and contains just a few records. That doesn't
matter too much, as we simply want to be able to initialise and use
analytics.GeoIPDB in the tests.

initialiseSystem doesn't return any error, so we have to check for nil
ourselves to be sure that the database was loaded correctly.

In preparation to fix #352 with a test.

Updates #352.

+++ gateway_test.go
@@ -8,6 +8,7 @@ import (
 	"net/http/httptest"
 	"net/url"
 	"os"
+	"path/filepath"
 	"reflect"
 	"strings"
 	"testing"
@@ -28,7 +29,12 @@ func TestMain(m *testing.M) {
 		panic(err)
 	}
 	config.EnableAnalytics = true
-	initialiseSystem(map[string]interface{}{})
+	config.AnalyticsConfig.EnableGeoIP = true
+	config.AnalyticsConfig.GeoIPDBLocation = filepath.Join("testdata", "MaxMind-DB-test-ipv4-24.mmdb")
+	initialiseSystem(nil)
+	if analytics.GeoIPDB == nil {
+		panic("GeoIPDB was not initialized")
+	}
 	exitCode := m.Run()
 
 	os.RemoveAll(config.AppPath)
